Рефакторинг

+++ src/main/java/dima/evseenko/amazon/ProductService.java
@@ -6,21 +6,22 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
-public class Amazon {
-    @Getter
+@Getter
+public class ProductService {
     private final Map<Long, Product> products = new HashMap<>();
 
     public void addProduct(Product product) {
-        validateProduct(product);
-
-        products.put(product.getId(), product);
+        if(Objects.nonNull(product)) {
+            products.put(product.getId(), product);
+        }
     }
 
     public void addProducts(List<Product> products) {
-        validateProducts(products);
-
-        products.forEach(this::addProduct);
+        if(Objects.nonNull(products)) {
+            products.forEach(this::addProduct);
+        }
     }
 
     public Product findProduct(Long id) {
@@ -32,21 +33,14 @@ public class Amazon {
     }
 
     public void deleteProduct(Product product) {
-        deleteProduct(product.getId());
-    }
-
-    public void deleteProducts(List<Product> products) {
-        products.forEach(this::deleteProduct);
+        if(Objects.nonNull(product)) {
+            deleteProduct(product.getId());
+        }
     }
 
     public List<Product> getProductsByCategory(Product.Category category) {
-        List<Product> productsByCategory = new ArrayList<>();
-        products.forEach((id, product) -> {
-            if(product.getCategory().equals(category)) {
-                productsByCategory.add(product);
-            }
-        });
-        return productsByCategory;
+        if(category == null) return null;
+        return groupProductsByCategory(this.products).get(category);
     }
 
     public void printProducts() {
@@ -55,22 +49,6 @@ public class Amazon {
 
     public void printGroupedProductsByCategory() {
         groupProductsByCategory(this.products).forEach((category, products) -> System.out.println(category+": "+products));
-    }
-
-    private void validateProduct(Product product) {
-        if(product == null)
-            throw new IllegalArgumentException("product is null");
-        if(product.getId() == null)
-            throw new IllegalArgumentException("product id is null");
-        if(product.getName() == null)
-            throw new IllegalArgumentException("product name is null");
-        if(product.getCategory() == null)
-            throw new IllegalArgumentException("product category is null");
-    }
-
-    private void validateProducts(List<Product> products) {
-        if(products == null)
-            throw new IllegalArgumentException("products is null");
     }
 
     public static Map<Product.Category, List<Product>> groupProductsByCategory(Map<Long, Product> products) {
